+++ examples/ek1100.rs
@@ -15,10 +15,7 @@
 //! ```
 
 use env_logger::Env;
-use ethercrab::{
-    error::Error, std::tx_rx_task, Client, ClientConfig, PduStorage, SlaveGroup, SlaveState,
-    Timeouts,
-};
+use ethercrab::{error::Error, std::tx_rx_task, Client, ClientConfig, PduStorage, Timeouts};
 use std::{sync::Arc, time::Duration};
 use tokio::time::MissedTickBehavior;
 
@@ -62,35 +59,28 @@ async fn main() -> Result<(), Error> {
     tokio::spawn(tx_rx_task(&interface, tx, rx).expect("spawn TX/RX task"));
 
     let mut group = client
-        .init_single_group::<MAX_SLAVES, PDI_LEN>(SlaveGroup::new(|slave| {
-            Box::pin(async {
-                // Special case: if an EL3004 module is discovered, it needs some specific config during
-                // init to function properly
-                if slave.name() == "EL3004" {
-                    log::info!("Found EL3004. Configuring...");
-
-                    slave.sdo_write(0x1c12, 0, 0u8).await?;
-                    slave.sdo_write(0x1c13, 0, 0u8).await?;
-
-                    slave.sdo_write(0x1c13, 1, 0x1a00u16).await?;
-                    slave.sdo_write(0x1c13, 2, 0x1a02u16).await?;
-                    slave.sdo_write(0x1c13, 3, 0x1a04u16).await?;
-                    slave.sdo_write(0x1c13, 4, 0x1a06u16).await?;
-                    slave.sdo_write(0x1c13, 0, 4u8).await?;
-                }
-
-                Ok(())
-            })
-        }))
+        .init_single_group::<MAX_SLAVES, PDI_LEN>()
         .await
         .expect("Init");
 
     log::info!("Discovered {} slaves", group.len());
 
-    client
-        .request_slave_state(SlaveState::Op)
-        .await
-        .expect("OP");
+    for slave in group.iter(&client) {
+        if slave.name() == "EL3004" {
+            log::info!("Found EL3004. Configuring...");
+
+            slave.sdo_write(0x1c12, 0, 0u8).await?;
+            slave.sdo_write(0x1c13, 0, 0u8).await?;
+
+            slave.sdo_write(0x1c13, 1, 0x1a00u16).await?;
+            slave.sdo_write(0x1c13, 2, 0x1a02u16).await?;
+            slave.sdo_write(0x1c13, 3, 0x1a04u16).await?;
+            slave.sdo_write(0x1c13, 4, 0x1a06u16).await?;
+            slave.sdo_write(0x1c13, 0, 4u8).await?;
+        }
+    }
+
+    let mut group = group.into_op(&client).await.expect("PRE-OP -> OP");
 
     for slave in group.iter(&client) {
         let (i, o) = slave.io_raw();