+++ tests/replay-ek1100-el2828-el2889.rs
@@ -3,7 +3,10 @@
 mod util;
 
 use env_logger::Env;
-use ethercrab::{error::Error, Client, ClientConfig, PduStorage, SlaveGroup, SlaveState, Timeouts};
+use ethercrab::{
+    error::Error, slave_group, Client, ClientConfig, PduStorage, SlaveGroup, SlaveGroupState,
+    Timeouts,
+};
 use std::time::Duration;
 use tokio::time::MissedTickBehavior;
 
@@ -16,9 +19,9 @@ struct Groups {
     /// EL2889 and EK1100. 2 items, 2 bytes of PDI for 16 output bits.
     ///
     /// We'll keep the EK1100 in here as it has no PDI but still needs to live somewhere.
-    slow_outputs: SlaveGroup<2, 2>,
+    slow_outputs: SlaveGroup<2, 2, slave_group::PreOp>,
     /// EL2828. 1 item, 1 byte of PDI for 8 output bits.
-    fast_outputs: SlaveGroup<1, 1>,
+    fast_outputs: SlaveGroup<1, 1, slave_group::PreOp>,
 }
 
 #[tokio::test]
@@ -43,7 +46,7 @@ async fn replay_ek1100_el2828_el2889() -> Result<(), Error> {
 
     // Read configurations from slave EEPROMs and configure devices.
     let groups = client
-        .init::<MAX_SLAVES, _>(Groups::default(), |groups, slave| match slave.name() {
+        .init::<MAX_SLAVES, _>(|groups: &Groups, slave| match slave.name() {
             "EL2889" | "EK1100" => Ok(&groups.slow_outputs),
             "EL2828" => Ok(&groups.fast_outputs),
             _ => Err(Error::UnknownSlave),
@@ -53,13 +56,11 @@ async fn replay_ek1100_el2828_el2889() -> Result<(), Error> {
 
     let Groups {
         slow_outputs,
-        mut fast_outputs,
+        fast_outputs,
     } = groups;
 
-    client
-        .request_slave_state(SlaveState::Op)
-        .await
-        .expect("OP");
+    let slow_outputs = slow_outputs.into_op(&client).await.expect("Slow into OP");
+    let mut fast_outputs = fast_outputs.into_op(&client).await.expect("Fast into OP");
 
     let mut slow_cycle_time = tokio::time::interval(Duration::from_millis(10));
     slow_cycle_time.set_missed_tick_behavior(MissedTickBehavior::Skip);